requirements.txt
asgiref==3.3.4
Django==3.2.4
django-ckeditor==6.1.0
django-filter==2.4.0
django-jazzmin==2.4.7
django-js-asset==1.2.2
django-phonenumber-field==5.2.0
django-satchmo==0.0.1
django-smart-selects==1.5.9
django-widget-tweaks==1.4.8
phonenumbers==8.12.31
Pillow==8.2.0
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
